requirements.txt
category_encoders==2.5.1.post0
numpy==1.23.5
optuna==3.0.5
pandas==1.3.5
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
rtdl==0.0.13
scikit_learn==1.0.2
scipy==1.8.1
torch==1.9.0+cu111
tqdm==4.64.1
typing_extensions==4.8.0
